requirements.txt
playwright==1.40.0
flask==3.0.0
gspread==5.12.0
oauth2client==4.1.3
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
